emoji ligatures should use "ccmp" rather than "liga"
Currently, if letter-spacing is applied in Firefox, the emoji ligatures such as regional-indicator flags or family groups will decompose, because Firefox automatically disables standard ligature features when letter-spacing (or inter-character justification) is in effect.

But unlike an "fi" ligature, we don't want 🇸🇪 or 👩‍👩‍👧‍👧 to split into its components just because letter-spacing is used. So the glyph composition feature "ccmp" is a better choice than "liga" for the OpenType rules here.